Move entities into model package, can split them out later as we add logic
