+++ src/main/java/uk/gov/mca/beacons/service/model/BeaconPerson.java
@@ -1,4 +1,4 @@
-package uk.gov.mca.beacons.service.beacon_person;
+package uk.gov.mca.beacons.service.model;
 
 import java.util.UUID;
 import javax.persistence.Entity;
