+++ src/main/java/uk/gov/mca/beacons/service/model/BeaconPersonRepository.java
@@ -1,8 +1,9 @@
-package uk.gov.mca.beacons.service.beacon;
+package uk.gov.mca.beacons.service.model;
 
 import java.util.UUID;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BeaconRepository extends CrudRepository<Beacon, UUID> {}
+public interface BeaconPersonRepository
+  extends CrudRepository<BeaconPerson, UUID> {}
