+++ src/main/java/uk/gov/mca/beacons/service/model/BeaconRepository.java
@@ -1,9 +1,8 @@
-package uk.gov.mca.beacons.service.beacon_person;
+package uk.gov.mca.beacons.service.model;
 
 import java.util.UUID;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BeaconPersonRepository
-  extends CrudRepository<BeaconPerson, UUID> {}
+public interface BeaconRepository extends CrudRepository<Beacon, UUID> {}
