+++ src/main/java/uk/gov/mca/beacons/service/model/PersonRepository.java
@@ -1,8 +1,8 @@
-package uk.gov.mca.beacons.service.beacon_use;
+package uk.gov.mca.beacons.service.model;
 
 import java.util.UUID;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BeaconUseRepository extends CrudRepository<BeaconUse, UUID> {}
+public interface PersonRepository extends CrudRepository<Person, UUID> {}
